+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Categories for each release: Added, Changed, Deprecated, Removed, Fixed, Securit
 
 ## Unreleased
 
+### Added
+
+* Display a security warning when downloading or generating download URLs for files flagged as malicious
+
 ## [409.0] - beta
 
 * No significant changes